double precision of nbody is added

+++ examples/oclNbody/inc/oclBodySystemOpenclLaunch_dp.h
@@ -25,7 +25,7 @@
     #include <CL/opencl.h>
 #endif 
 
-#include "oclBodySystemOpencl.h"
+#include "oclBodySystemOpencl_dp.h"
 
 int  CreateProgramAndKernel(cl_context ctx, cl_device_id* cdDevices, const char* kernel_name, cl_kernel* kernel, bool bDouble);
 void AllocateNBodyArrays(cl_context ctx, cl_mem* vel, int numBodies, int dFlag);
@@ -37,11 +37,11 @@ void IntegrateNbodySystem(cl_command_queue cqCommandQueue,
                           cl_mem newPos, cl_mem newVel,
                           cl_mem oldPos, cl_mem oldVel,
                           cl_mem pboCLOldPos, cl_mem pboCLNewPos,
-                          float deltaTime, float damping, float softSq,
+                          double deltaTime, double damping, double softSq,
                           int numBodies, int p, int q,
                           int bUsePBO, bool bDouble);
-void CopyArrayFromDevice(cl_command_queue cmdq, float *host, cl_mem device, cl_mem pboCL, int numBodies, bool bDouble);
-void CopyArrayToDevice(cl_command_queue cmdq, cl_mem device, const float *host, int numBodies, bool bDouble);
+void CopyArrayFromDevice(cl_command_queue cmdq, double *host, cl_mem device, cl_mem pboCL, int numBodies, bool bDouble);
+void CopyArrayToDevice(cl_command_queue cmdq, cl_mem device, const double *host, int numBodies, bool bDouble);
 cl_mem RegisterGLBufferObject(cl_context ctx, unsigned int pboGL);
 void UnregisterGLBufferObject(cl_mem pboCL);
 void ThreadSync(cl_command_queue cmdq);
